backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
pytest==7.4.3
pytest-asyncio==0.21.1
gunicorn==21.2.0
geopy==2.4.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
transformers==4.35.2
python-multipart==0.0.6
twilio==8.10.0
stripe==7.4.0
email-validator==2.1.0
pydantic-extra-types==2.10.6